feature(frontend&backend): implement custom server starting, stopping, logs & cleaned up sidebar elements

+++ src-tauri/src/custom_servers/providers/paper.rs
@@ -25,13 +25,15 @@ impl PaperProvider {
 
     pub async fn download_server_jar<F>(custom_server: &CustomServer, on_progress: F) -> Result<()> where F : Fn(u64, u64) {
         let path = LAUNCHER_DIRECTORY.data_dir().join("custom_servers").join(&custom_server.id);
-        fs::create_dir_all(&path).await?;
-        let mut build_version = Self::get_all_build_versions(&custom_server.mc_version).await?.builds;
-        build_version.reverse();
-        let latest_build = build_version.first().unwrap();
-        let url = format!("{}/versions/{}/builds/{}/downloads/{}", PAPER_API_BASE, &custom_server.mc_version, custom_server.loader_version.clone().unwrap_or_default(), format!("server-{}-{}.jar", custom_server.mc_version, latest_build.build));
-        let content = download_file(&url, on_progress).await?;
-        let _ = fs::write(path.join("server.jar"), content).await.map_err(|e| e);
+        if !path.join("server.jar").exists() {
+            fs::create_dir_all(&path).await?;
+            let mut build_version = Self::get_all_build_versions(&custom_server.mc_version).await?.builds;
+            build_version.reverse();
+            let latest_build = build_version.first().unwrap();
+            let url = format!("{}/versions/{}/builds/{}/downloads/{}", PAPER_API_BASE, &custom_server.mc_version, custom_server.loader_version.clone().unwrap_or_default(), format!("server-{}-{}.jar", custom_server.mc_version, latest_build.build));
+            let content = download_file(&url, on_progress).await?;
+            let _ = fs::write(path.join("server.jar"), content).await.map_err(|e| e);
+        }
         Ok(())
     }
